final clean up
